% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{cromwell_base}
\alias{cromwell_base}
\title{Fetch the base URL and port for cromwell server}
\usage{
cromwell_base()
}
\description{
The Cromwell server presents a RESTFul API. The base URL is of the form:
`http://EXAMPLE.COM:PORT`. The current approach to changing that
url is to set an option, `cromwell_base` to a valid URL (without trailing slash).
This URL will then be used throughout the `cRomwell` package.  If no option is set,
the server is assumed to be running at `http://localhost:8000`.
}
\examples{
cromwell_base()

# set a bogus host
options('cromwell_base' = 'http://example.com:8111')
cromwell_base()

# and set back to NULL to get the default behavior
options('cromwell_base' = NULL)
cromwell_base()
}

