% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{cromwell_GET}
\alias{cromwell_GET}
\title{Perform a GET request to cromwell server}
\usage{
cromwell_GET(path, query = NULL, ...)
}
\arguments{
\item{path}{The path part of the URL}

\item{query}{Any query terms as a named character vector}

\item{...}{passed directly to httr `GET` (for including `timeouts`,
`handles`, etc.)}
}
\description{
See the docmentation at
\href{https://github.com/broadinstitute/cromwell#rest-api}{the
cromwell github site} for details. Generally, this is not meant to
be called by the end user. Rather, use the endpoint-specific
functions. See \code{\link{cromwellBase}} for details of setting
the base URL and port.
}
\keyword{internal}
