% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{cromwell_POST}
\alias{cromwell_POST}
\title{Perform a POST request to cromwell server}
\usage{
cromwell_POST(path, body, ...)
}
\arguments{
\item{path}{The path part of the URL}

\item{body}{A list that will become the multipart form that is passed as the request body}

\item{...}{passed directly to httr `POST` (for including `timeouts`, `handles`, etc.)}
}
\description{
See the docmentation at
\href{https://github.com/broadinstitute/cromwell#rest-api}{the
cromwell github site} for details. Generally, this is not meant to
be called by the end user. Rather, use the endpoint-specific
functions.
}
\seealso{
\code{\link{cromwellBatch}}
}

