% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{cromwellAbort}
\alias{cromwellAbort}
\title{Abort a cromwell job}
\usage{
cromwellAbort(id, ...)
}
\arguments{
\item{id}{A cromwell id as a string}

\item{...}{passed directly to httr `POST` (for including `timeouts`, `handles`, etc.)}
}
\description{
Abort a cromwell job
}
\examples{
#cromwellAbord('ID')
}
