% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{cromwellBase}
\alias{cromwellBase}
\alias{setCromwellBase}
\title{Set and get base URL and port for cromwell server}
\usage{
cromwellBase()

setCromwellBase(base_url = NULL)
}
\arguments{
\item{base_url}{character(1) specifying the url and port of the
cromwell server}
}
\value{
A character(1) base url (default \code{http://localhost:8000})
}
\description{
The Cromwell server presents a RESTFul API. The base URL is of the form:
`http://EXAMPLE.COM:PORT`. The current approach to changing that
url is to set an option, `cromwellBase` to a valid URL (without trailing slash).
This URL will then be used throughout the `cRomwell` package.  If no option is set,
the server is assumed to be running at `http://localhost:8000`.
}
\examples{
cromwellBase()

# set a bogus host
setCromwellBase('http://example.com:8111')
cromwellBase()

# and set back to NULL to get the default behavior
setCromwellBase()
cromwellBase()

}
