% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{cromwellMetadata}
\alias{cromwellMetadata}
\title{Get metadata associated with one or more workflow ids}
\usage{
cromwellMetadata(id, ...)
}
\arguments{
\item{id}{A cromwell id as a string}

\item{...}{passed directly to httr `GET` (for including `timeouts`, `handles`, etc.)}
}
\value{
a list of metadata lists
}
\description{
Get metadata associated with one or more workflow ids
}
\examples{
#cromwellQuery(ids=c('1','2','abc'))
}

