% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{cromwellOutputs}
\alias{cromwellOutputs}
\title{Get output paths associated with one or more workflow ids}
\usage{
cromwellOutputs(id)
}
\arguments{
\item{id}{}
}
\value{
a list of output lists
}
\description{
Get output paths associated with one or more workflow ids
}
\examples{
#cromwellOutputs(ids)
}

