% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{cromwellQuery}
\alias{cromwellQuery}
\title{Get the info about cromwell workflows}
\usage{
cromwellQuery(terms = NULL)
}
\arguments{
\item{terms}{}
}
\value{
a data.frame of query results
}
\description{
Get the info about cromwell workflows
}
\examples{
#cromwellQuery(terms='status=Succeeded')
}

