% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{.cromwell_process_response}
\alias{.cromwell_process_response}
\title{Check cromwell response}
\usage{
.cromwell_process_response(resp)
}
\arguments{
\item{resp}{a \code{\link{response}} object}
}
\value{
a simple list that includes the actual `content` and the complete `response` object.
}
\description{
Check cromwell response
}
