% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{getCromwellJar}
\alias{getCromwellJar}
\title{Get the cromwell JAR file from github}
\usage{
getCromwellJar(cromwell_version = "23", destfile = "cromwell.jar")
}
\arguments{
\item{cromwell_version}{string representing the version number}

\item{destfile}{string The full path to the cromwell jar file location on the local system}
}
\value{
destfile location [invisibly]
}
\description{
This function simply downloads the cromwell JAR file
and puts it in the destfile location. The JAR file is picked up from
\url{https://github.com/broadinstitute/cromwell/releases}.
}
\examples{
version = '24'
tmpfile = file.path(tempdir(),'cromwell.jar')
fp = getCromwellJar(cromwell_version = version, destfile = tmpfile)
fp
unlink(fp)

}

