% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{getWdltoolJar}
\alias{getWdltoolJar}
\title{Utility to fetch the wdltool JAR file}
\usage{
getWdltoolJar(wdltool_version, destfile = "wdltool.jar")
}
\arguments{
\item{wdltool_version}{string representing the version number}

\item{destfile}{string The full path to the wdltool jar file
location on the local system}
}
\value{
destfile location [invisibly]
}
\description{
This function simply downloads the wdltool JAR file and puts it in
the destfile location. The JAR file is picked up from
\url{https://github.com/broadinstitute/wdltool/releases}.
}
\examples{
version = '0.8'
tmpfile = file.path(tempdir(),'wdltool.jar')
fp = getWdltoolJar(wdltool_version = version)
fp
unlink(fp)

}
\seealso{
See lots of details at \url{https://github.com/broadinstitute/wdltool}.
}
