% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\docType{class}
\name{wdlrunr-class}
\alias{wdlrunr-class}
\title{Cromwell reference class}
\description{
Cromwell reference class
}
\section{Methods}{

\describe{
\item{\code{initialize(host = "localhost", port = 8000)}}{This is test documentation}
}}

