% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\docType{data}
\name{wdlrunr}
\alias{wdlrunr}
\title{Cromwell reference class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
wdlrunr
}
\description{
Cromwell reference class
}
\keyword{datasets}
