% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdltool.R
\name{wdltoolInputs}
\alias{wdltoolInputs}
\title{return inputs from wdltool}
\usage{
wdltoolInputs(wdlfile, wdltoolJar = Sys.getenv("WDLTOOL_JAR", "wdltool.jar"))
}
\arguments{
\item{wdlfile}{The filename of a WDL file from which to get inputs.}

\item{wdltoolJar}{The filename of the wdltool jar file; if not
specified, the environment variable \code{WDLTOOL_JAR} is
consulted.}
}
\value{
a 2-column data.frame with the input \code{name} and the
    input \code{type} as the two columns.
}
\description{
This little function just runs the command-line wdltool inputs on a
wdl file and returns a data.frame of input names and input types.
}
\examples{
version = '0.8'
tmpfile = file.path(tempdir(),'wdltool.jar')
fp = getWdltoolJar(wdltool_version = version,
                   destfile = tmpfile)
wdltoolInputs(system.file(package='cRomwell','wdl/fastqc_task.wdl'))

}
