% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{cromwellBackends}
\alias{cromwellBackends}
\title{List available backends for a cromwell endpoint}
\usage{
cromwellBackends(...)
}
\arguments{
\item{...}{passed directly to httr `GET` (for including `timeouts`, `handles`, etc.)}
}
\value{
a list that includes backend details
}
\description{
This endpoint returns a list of the backends supported by the
server as well as the default backend.
}
\examples{
#cromwellBackends()
}
\references{
\url{https://github.com/broadinstitute/cromwell#get-apiworkflowsversionbackends}
}
