% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{cromwellBase}
\alias{cromwellBase}
\title{Fetch the base URL and port for cromwell server}
\usage{
cromwellBase()
}
\description{
The Cromwell server presents a RESTFul API. The base URL is of the form:
`http://EXAMPLE.COM:PORT`. The current approach to changing that
url is to set an option, `cromwellBase` to a valid URL (without trailing slash).
This URL will then be used throughout the `cRomwell` package.  If no option is set,
the server is assumed to be running at `http://localhost:8000`.
}
\examples{
cromwellBase()

# set a bogus host
options('cromwellBase' = 'http://example.com:8111')
cromwellBase()

# and set back to NULL to get the default behavior
options('cromwellBase' = NULL)
cromwellBase()

}
