% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{cromwellBatch}
\alias{cromwellBatch}
\title{Submit a cromwell batch job}
\usage{
cromwellBatch(wdlSource, workflowInputs, workflowOptions = NULL,
  timeout = 120, ...)
}
\arguments{
\item{wdlSource}{A \code{list}, a JSON string (as a \code{character} vector of length 1,
or an \code{\link[httr]{upload_file}} object. See details below.}

\item{workflowInputs}{A \code{list}, a JSON string (as a \code{character} vector of length 1,
or an \code{\link[httr]{upload_file}} object. See details below.}

\item{workflowOptions}{A \code{list}, a JSON string (as a \code{character} vector of length 1,
or an \code{\link[httr]{upload_file}} object. See details below.}

\item{timeout}{The number of seconds to wait for a response. Batch jobs can take
quite some time for cromwell to process, so this will typically need to be set
to a large value to allow for a completed response.}

\item{...}{passed directly to httr `POST` (for including `timeouts`, `handles`, etc.)}
}
\description{
This function submits a set of one or more inputs to cromwell. It is much more efficient
than submitting a single job at a time.  See
\href{https://github.com/broadinstitute/cromwell#post-apiworkflowsversionbatch}{the cromwell \code{batch} API documentation} for details.
}
\details{
abc details
}

