% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{cromwellLogs}
\alias{cromwellLogs}
\title{Get output paths associated with one or more workflow ids}
\usage{
cromwellLogs(id)
}
\arguments{
\item{id}{}
}
\value{
a list of logfile lists
}
\description{
Get output paths associated with one or more workflow ids
}

