% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{cromwellMetadata}
\alias{cromwellMetadata}
\title{Get metadata associated with one or more workflow ids}
\usage{
cromwellMetadata(ids, query = NULL, ...)
}
\arguments{
\item{ids}{A character() vector of cromwell IDs, typically returned
from a batch submission or from a call to \code{\link{cromwellQuery}}.}

\item{...}{passed directly to httr `GET` (for including `timeouts`,
`handles`, etc.)}
}
\value{
a list of metadata lists
}
\description{
This endpoint returns a superset of the data from
#get-workflowsversionidlogs in essentially the same format
(i.e. shards are accounted for by an array of maps, in the same
order as the shards). In addition to shards, every attempt that was
made for this call will have its own object as well, in the same
order as the attempts. Workflow metadata includes submission,
start, and end datetimes, as well as status, inputs and
outputs. Call-level metadata includes inputs, outputs, start and
end datetime, backend-specific job id, return code, stdout and
stderr. Date formats are ISO with milliseconds.
}
\examples{
\dontrun{
res = cromwellQuery(terms=c(status='Succeeded',name='taskName'))
head(res)
metalist = cromwellMetadata(res$id)
str(metalist,list.len=5)
}

}
\references{
\url{https://github.com/broadinstitute/cromwell#get-apiworkflowsversionidmetadata}
}
