% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{cromwellQuery}
\alias{cromwellQuery}
\title{Get the info about cromwell workflows}
\usage{
cromwellQuery(terms = NULL, ...)
}
\arguments{
\item{terms}{terms about the workflow}

\item{...}{passed directly to httr `GET` (for including `timeouts`, `handles`, etc.)}
}
\value{
a data.frame of query results
}
\description{
Get the info about cromwell workflows
}
\examples{
#cromwellQuery(terms=c(status='Succeeded',name='taskName'))
}

