% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{cromwellQuery}
\alias{cromwellQuery}
\title{Get the info about cromwell workflows}
\usage{
cromwellQuery(terms = NULL, ...)
}
\arguments{
\item{terms}{terms about the workflow}

\item{...}{passed directly to httr `GET` (for including `timeouts`, `handles`, etc.)}
}
\value{
a data.frame of query results
}
\description{
Each of the following terms can be specified one or more times. Simply create a named list
or named character vector.
\describe{
  \item{name}{The name of a job; may be specified more than once}
  \item{status}{one of Succeeded, Failed, Running}
  \item{id}{an id of a cromwell job}
  \item{start}{a timestamp of the form "2015-11-01T07:45:52.000-05:00", including mandatory offset}
  \item{end}{a timestamp of the form "2015-11-01T07:45:52.000-05:00", including mandatory offset}
  \item{page}{if paging is used, what page to select}
  \item{pagesize}{if paging is used, how many records per page}
}
}
\examples{
\dontrun{
res = cromwellQuery(terms=c(status='Succeeded',name='taskName'))
head(res)
}

}
