% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{cromwellQuery}
\alias{cromwellQuery}
\title{Get the info about cromwell workflows}
\usage{
cromwellQuery(name = NULL, id = NULL, status = NULL, start = NULL,
  end = NULL, label = NULL, ...)
}
\arguments{
\item{name}{character vector of workflow names.}

\item{id}{character vector of workflow IDs.}

\item{status}{character vector of workflow status values. The available
statuses are: Submitted, Running, Aborting, Aborted, Failed, and Succeeded.}

\item{start}{a \code{Date} object specifying the workflow start time. Only workflows
started after this are returned.}

\item{end}{a \code{Date} object specifying the workflow end time. Only workflows
that completed after this are returned.}

\item{label}{character vector of workflow labels.}

\item{...}{passed directly to httr `GET` (for including `timeouts`,
`handles`, etc.)}
}
\value{
a data.frame of query results
}
\description{
Each of the following terms can be specified one or more
times. Simply create a named list or named character vector.
}
\details{
\describe{
  \item{name}{The name of a job; may be specified more than once}
  \item{status}{one of Succeeded, Failed, Running}
  \item{id}{an id of a cromwell job}
  \item{start}{a timestamp of the form "2015-11-01T07:45:52.000-05:00", including mandatory offset}
  \item{end}{a timestamp of the form "2015-11-01T07:45:52.000-05:00", including mandatory offset}
  \item{page}{if paging is used, what page to select}
  \item{pagesize}{if paging is used, how many records per page}
}
}
\examples{
\dontrun{
res = cromwellQuery(status='Succeeded')
head(res)
}

}
