% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{cromwellStats}
\alias{cromwellStats}
\title{Get current statistics for cromwell endpoint}
\usage{
cromwellStats(...)
}
\arguments{
\item{...}{passed directly to httr `GET` (for including `timeouts`, `handles`, etc.)}
}
\value{
a list
}
\description{
Get current statistics for cromwell endpoint
}
\examples{
#cromwellStats()
}

