% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{getCromwellJar}
\alias{getCromwellJar}
\title{Utility to fetch the cromwell JAR file}
\usage{
getCromwellJar(cromwell_version, destfile = file.path(tempdir(),
  "cromwell.jar"))
}
\arguments{
\item{cromwell_version}{string representing the version number}

\item{destfile}{string The full path to the cromwell jar file location on the local system}
}
\value{
destfile location [invisibly]
}
\description{
The purpose of this R package is to interact with the
Broad Cromwell execution engine.
The Cromwell server is contained in a JAVA JAR file. This
function simply downloads the cromwell JAR file
and puts it in the destfile location. The JAR file is picked up from
\url{https://github.com/broadinstitute/cromwell/releases}.
}
\examples{
version = '24'
tmpfile = file.path(tempdir(),'cromwell.jar')
fp = getCromwellJar(cromwell_version = version)
fp
unlink(fp)

}
\seealso{
See lots of details at \url{https://github.com/broadinstitute/cromwell}.
}

